% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alabama.R
\docType{data}
\name{alabama}
\alias{alabama}
\title{American community survey data from the state of Alabama.}
\format{A data frame with 46847 rows and 7 variables:
\describe{
  \item{Age}{age of the individual, in years}
  \item{Medicare}{an indicator for whether the individual receives Medicare benefits}
  \item{Medicaid}{an indicator for whether the individual receives Medicaid benefits}
  \item{Disability}{an indicator for whether the individual has a disability}
  \item{Educ}{level of education from 0 (under 3 years old) to 24 (doctorate)}
  \item{Income}{annual income of the individual, in US dollars}
  \item{HrsWorked}{number of hours worked per week}
}}
\usage{
alabama
}
\description{
A public-use dataset containing information about a 1 percent subsample of Alabama
residents in the American community survey.
}
\keyword{datasets}
