% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_transformations.R
\name{applyTransformations}
\alias{applyTransformations}
\title{Apply a list of transformations across a matrix.}
\usage{
applyTransformations(X, transformations, indices = NULL)
}
\arguments{
\item{X}{A matrix.}

\item{transformations}{A list of functions to apply to corresponding columns.}

\item{indices}{The columns of the matrix for which to apply the transformation.}
}
\value{
A matrix (or vector) of transformed values.
}
\description{
Apply a list of transformations across a matrix.
}
