% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R, R/midamix.R
\docType{package}
\name{midamix}
\alias{midamix}
\alias{midamix-package}
\title{Multiple imputation for data analysis via mixtures}
\usage{
midamix(data, imputations = 10, max_clusters = 15, n_iter = 1000,
  burnin = 100, validator = NULL, transformations = NULL,
  seed = NA)
}
\arguments{
\item{data}{A tibble, data frame, or matrix, consisting of numeric, integer,
or ordered factor columns.}

\item{imputations}{The number of imputed datasets to return, defaults to 10.}

\item{max_clusters}{The maximum number of clusters for the mixture model.}

\item{n_iter}{Number of iterations for the MCMC sampler.}

\item{burnin}{Number of iterations for initial burn-in period.}

\item{validator}{A function that indicates the validity of imputations.  If \code{NULL},
the data are assumed unconstrained.}

\item{transformations}{Transformations of response variables to latent scale, defaults
to empirical CDFs}

\item{seed}{Random seed.}
}
\value{
A \code{tibble} consisting of multiply imputed data sets.
}
\description{
Uses a Dirichlet Process Mixture Transformation Model to sample missing data values
from the corresponding posterior distribution.

Multiple imputation for data analysis via mixtures
}
