% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nflcombine.R
\docType{data}
\name{nfl_combine}
\alias{nfl_combine}
\title{National Football League Scouting Combine Data}
\format{A data frame with 10502 rows and 13 variables:
\describe{
  \item{year}{year}
  \item{name}{player name}
  \item{college}{university the player attended}
  \item{position}{primary position}
  \item{height}{height of the player, in inches}
  \item{weight}{weight of the player, in pounds}
  \item{wonderlic}{score on the wonderlic intelligence test, ranging from 0-50}
  \item{forty_yard_dash}{time taken to run the forty yard dash, in seconds}
  \item{bench_press}{number of 225 lb repetitions player could bench press}
  \item{vertical_jump}{height of maximum vertical jump, in inches}
  \item{broad_jump}{length of maximum player broad jump, in inches}
  \item{shuttle}{time taken to run the twenty yard shuttle, in seconds}
  \item{three_cone}{time taken to run the three-cone drill, in seconds}
}}
\usage{
nfl_combine
}
\description{
A compilation of drill results all players who attended the NFL
scouting combine from 1987 to 2019.  Some of the measurements come from the player's
pro day.
}
\keyword{datasets}
