% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_inferences.R
\name{pool_inferences}
\alias{pool_inferences}
\title{Combine Multiple Imputations}
\usage{
pool_inferences(model_fits, by_significance = FALSE)
}
\arguments{
\item{model_fits}{An object returned by the function \code{fit_model}}

\item{by_significance}{Should the p-values be ordered in ascending order?}
}
\value{
A tibble with inferences obtained by multiple imputation combining rules.
}
\description{
Pools inferences for regression coefficients for linear models and generalized linear
models
}
