% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mixtures.R
\name{rMixtureNormalTransform}
\alias{rMixtureNormalTransform}
\title{Generates data from a transformed mixture of multivariate normal distributions.}
\usage{
rMixtureNormalTransform(n, mean = matrix(0, nrow = dim(sigma)[1], ncol =
  num_clusters), sigma = array(rep(diag(nrow(mean)), num_clusters), dim =
  c(nrow(mean), nrow(mean), num_clusters)), num_clusters = 1,
  prob = rep(1, num_clusters)/num_clusters,
  transformations = rep(list(function(y) y), nrow(mean)),
  method = c("eigen", "svd", "chol"))
}
\arguments{
\item{n}{The number of samples to draw.}

\item{mean}{A matrix corresponding to the means of each cluster.}

\item{sigma}{An array with the covariance matrices of each cluster.}

\item{num_clusters}{The number of clusters in the mixture.}

\item{prob}{Probabilities of cluster membership.}

\item{transformations}{A list of functions to apply to corresponding columns.}

\item{method}{Method used for generating from a multivariate normal distribution.}
}
\value{
A matrix of samples from a transformed mixture of multivariate normal distributions.
}
\description{
Generates data from a transformed mixture of multivariate normal distributions.
}
