% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rmvnorm}
\alias{rmvnorm}
\alias{dmvnorm}
\alias{dmvnorm_prec}
\title{Multivariate normal distribution}
\usage{
rmvnorm(n, mean = rep(0, nrow(sigma)), sigma = diag(length(mean)),
  method = c("eigen", "svd", "chol"), pre0.9_9994 = FALSE)

dmvnorm(x, mean = rep(0, p), sigma = diag(p), log = FALSE)

dmvnorm_prec(x, mean = rep(0, p), prec = diag(p), log = FALSE)
}
\arguments{
\item{n}{The number of observations}

\item{mean}{Mean vector, default is 0}

\item{sigma}{Covariance matrix, default is identity.}

\item{method}{string specifying the matrix decomposition used to determine the
matrix root of sigma}

\item{pre0.9_9994}{logical; version of rmvnorm used, inherited from mvtnorm package}

\item{x}{A vector or matrix of quantiles.  If x is a matrix, each row is taken to
be a quantile}

\item{log}{logical; if \code{TRUE}, log-densities are returned}

\item{prec}{Precision matrix, default is identity}
}
\description{
Random number generator and density for the multivariate normal distribution with mean equal
to \code{mean} and covariance matrix \code{sigma}
}
