% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rtruncnorm}
\alias{rtruncnorm}
\alias{dtruncnorm}
\alias{ptruncnorm}
\alias{qtruncnorm}
\title{Truncated normal distribution}
\usage{
rtruncnorm(n, a = -Inf, b = Inf, mean = 0, sd = 1)

dtruncnorm(y, a = -Inf, b = Inf, mean = 0, sd = 1, log = F)

ptruncnorm(y, a = -Inf, b = Inf, mean = 0, sd = 1)

qtruncnorm(u, a = -Inf, b = Inf, mean = 0, sd = 1)
}
\arguments{
\item{n}{Number of observations}

\item{a}{Lower bound}

\item{b}{Upper bound}

\item{mean}{Mean of the unconstrained distribution}

\item{sd}{Standard deviation of the unconstrained distribution}

\item{y}{A vector of data values}

\item{log}{logical; should the log-density be returned?}

\item{u}{A vector of quantiles}
}
\description{
Calculates densities and cdfs of the truncated normal distribution.  Also can generate from the distribution.
}
