% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rwish}
\alias{rwish}
\alias{riwish}
\title{Random Wishart distributions}
\usage{
rwish(v, S)

riwish(v, S)
}
\arguments{
\item{v}{degrees of freedom}

\item{S}{inverse scale matrix}
}
\value{
A positive definite matrix with the same dimension as S
}
\description{
Randomly generates positive definite matrices from Wishart and Inverse-Wishart distributions
}
